nile: Enforce VINTF kernel requirements
This resolves the build error:
"Unable to determine UFFD GC flag because the kernel version is not available"

Enforcing VINTF kernel requirements allows the build system to properly
detect the kernel version and automatically determine the appropriate
setting for userfaultfd garbage collection (UFFD GC).

This is the recommended solution as specified by the build system itself.

Signed-off-by: Yehonatan5785 <yehonatan@mail.ru>